+++ server/app/Http/Controllers/CartItemController.php
@@ -5,9 +5,12 @@ namespace App\Http\Controllers;
 use App\Models\Cart_item;
 use App\Http\Requests\StoreCart_itemRequest;
 use App\Http\Requests\UpdateCart_itemRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class CartItemController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
@@ -22,6 +25,8 @@ class CartItemController extends Controller
      */
     public function store(StoreCart_itemRequest $request)
     {
+        $this->authorize('create', Cart_item::class);
+        
         $validated = $request->validated();
         $cartItem = Cart_item::create($validated);
         return response()->json($cartItem, 201);
@@ -32,6 +37,8 @@ class CartItemController extends Controller
      */
     public function show(Cart_item $cart_item)
     {
+        $this->authorize('view', $cart_item);
+        
         return response()->json($cart_item);
     }
 
@@ -40,6 +47,8 @@ class CartItemController extends Controller
      */
     public function update(UpdateCart_itemRequest $request, Cart_item $cart_item)
     {
+        $this->authorize('update', $cart_item);
+        
         $validated = $request->validated();
         $cart_item->update($validated);
         return response()->json($cart_item);
@@ -50,6 +59,8 @@ class CartItemController extends Controller
      */
     public function destroy(Cart_item $cart_item)
     {
+        $this->authorize('delete', $cart_item);
+        
         $cart_item->delete();
         return response()->json(['message' => 'Deleted successfully']);
     }
