+++ server/app/Http/Controllers/ParameterController.php
@@ -5,9 +5,12 @@ namespace App\Http\Controllers;
 use App\Models\Parameter;
 use App\Http\Requests\StoreParameterRequest;
 use App\Http\Requests\UpdateParameterRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ParameterController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
@@ -22,6 +25,8 @@ class ParameterController extends Controller
      */
     public function store(StoreParameterRequest $request)
     {
+        $this->authorize('create', Parameter::class);
+        
         $validated = $request->validated();
         $parameter = Parameter::create($validated);
         return response()->json($parameter, 201);
@@ -40,6 +45,8 @@ class ParameterController extends Controller
      */
     public function update(UpdateParameterRequest $request, Parameter $parameter)
     {
+        $this->authorize('update', $parameter);
+        
         $validated = $request->validated();
         $parameter->update($validated);
         return response()->json($parameter);
@@ -50,6 +57,8 @@ class ParameterController extends Controller
      */
     public function destroy(Parameter $parameter)
     {
+        $this->authorize('delete', $parameter);
+        
         $parameter->delete();
         return response()->json(['message' => 'Deleted successfully']);
     }
