+++ server/app/Http/Controllers/ProductParameterController.php
@@ -5,9 +5,12 @@ namespace App\Http\Controllers;
 use App\Models\ProductParameter;
 use App\Http\Requests\StoreProduct_parameterRequest;
 use App\Http\Requests\UpdateProduct_parameterRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ProductParameterController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
@@ -22,6 +25,8 @@ class ProductParameterController extends Controller
      */
     public function store(StoreProduct_parameterRequest $request)
     {
+        $this->authorize('create', ProductParameter::class);
+        
         $validated = $request->validated();
         $param = ProductParameter::create($validated);
         return response()->json($param, 201);
@@ -40,6 +45,8 @@ class ProductParameterController extends Controller
      */
     public function update(UpdateProduct_parameterRequest $request, ProductParameter $product_parameter)
     {
+        $this->authorize('update', $product_parameter);
+        
         $validated = $request->validated();
         $product_parameter->update($validated);
         return response()->json($product_parameter);
@@ -50,6 +57,8 @@ class ProductParameterController extends Controller
      */
     public function destroy(ProductParameter $product_parameter)
     {
+        $this->authorize('delete', $product_parameter);
+        
         $product_parameter->delete();
         return response()->json(['message' => 'Deleted successfully']);
     }
